Miglioramenti divisione file
