+++ gdt server/Tesi/public/js/templates/main-view-technosphere-flows.js
@@ -28,67 +28,7 @@ function creaViewRowTotalRequirements(element,num){
     `;
 }
 
-function creaViewTableTechnosphereFlows() {
-    return `
-        <table class="table table-striped table-bordered table-scrollabile">
-            <thead>
-                <tr>
-                    <th scope="col">#5.5</th>
-                    <th scope="col">Nome Provider</th>
-                    <th scope="col">Nome Flow</th>
-                    <th scope="col">Unità</th>
-                </tr>
-            </thead>
-            <tbody id="datiTabellaTechnosphereFlows">
-            </tbody>
-        </table>  
-    `;
-  }
-  
-  function creaViewRowTechnosphereFlows(element,num){
-    return `
-    <tr class="color-row-table">
-        <th scope="row">${num}</th>
-        <td>${element.provider.name}</td>
-        <td>${element.flow.name}</td>
-        <td>${element.flow.refUnit}</td>      
-    </tr>
-    `;
-  }
-  
-  function creaViewTableFinalDemand() {
-    return `
-        <table class="table table-striped table-bordered">
-            <thead>
-                <tr>
-                    <th scope="col">#5.5.2</th>
-                    <th scope="col">Nome Provider</th>
-                    <th scope="col">Categoria Provider</th>
-                    <th scope="col">Nome Flow</th>
-                    <th scope="col">Categoria Flow</th>
-                    <th scope="col">Quantità</th>
-                </tr>
-            </thead>
-            <tbody id="datiTabellaFinalDemand">
-            </tbody>
-        </table>  
-    `;
-  }
-  
-  function creaViewRowFinalDemand(element,num){
-    return `
-    <tr class="color-row-table">
-        <th scope="row">${num}</th>
-        <td>${element.techFlow.provider.name}</td>
-        <td>${element.techFlow.provider.category}</td>      
-        <td>${element.techFlow.flow.name}</td>
-        <td>${element.techFlow.flow.category}</td>
-        <td>${element.amount}</td>
-    </tr>
-    `;
-  }
-  
-  function creaViewTabletotalRequirementsOfFlows() {
+function creaViewTabletotalRequirementsOfFlows() {
     return `
         <table class="table table-striped table-bordered"">
             <thead>
@@ -253,8 +193,6 @@ function creaViewTableTechnosphereFlows() {
   }
 
   export { creaViewTableTotalRequirements,creaViewRowTotalRequirements,
-    creaViewTableTechnosphereFlows, creaViewRowTechnosphereFlows,
-    creaViewTableFinalDemand,creaViewRowFinalDemand,
     creaViewTabletotalRequirementsOfFlows,creaViewRowtotalRequirementsOfFlows,
     creaViewTableScalingFactors,creaViewRowScalingFactors,
     creaViewTableTotalityFactors,creaViewRowTotalityFactors,
